feat: adding a check for valid tag type
adding a check when loading, by comparing the type written in file and
type as it should be according to standard. if type is not as it should
be, the tag is not loaded into the dictionary.
added try/except around unpacks, so that if the basic read fails, that
tag is skipped